# SecureTransferBackend/SecureTransferBackend/Services/Transfers/Dtos/DecryptorKeyDto.cs
namespace SecureTransferBackend.Services.Transfers.Dtos;

public class DecryptorKeyDto
{
    public string EncryptedSymmetricKey { get; set; }
    public string PublicKeyUsedForEncryption { get; set; }
}